refactored login method into component, add describe block in test
