+++ cypress/integration/journalistCanLogin.feature.js
@@ -41,21 +41,23 @@ describe('Journalist authenticates', () => {
       cy.get('button#post').should('be.visible')
     })
   })
-    
-    it("unsuccessfully with invalid credentials", () => {
+
+  describe('unsuccessfully', () => {
+    it("with invalid credentials", () => {
       cy.route({
-      method: "POST",
-      url: "http://localhost:3000/api/auth/sign_in",
-      response: "fixture:unsuccessful_login.json",
-      headers: {
-        uid:"user@mail.com"
-      }
-    })
-    cy.get("#login-form").within(() => {
-      cy.get("#email").type("user@mail.com");
-      cy.get("#password").type("wrongpassword");
-      cy.get('Button').contains('Submit').click()
+        method: "POST",
+        url: "http://localhost:3000/api/auth/sign_in",
+        response: "fixture:unsuccessful_login.json",
+        headers: {
+          uid:"user@mail.com"
+        }
+      })
+      cy.get("#login-form").within(() => {
+        cy.get("#email").type("user@mail.com");
+        cy.get("#password").type("wrongpassword");
+        cy.get('Button').contains('Submit').click()
+      });
+      cy.get("#error-message").should("contain", "Invalid login credentials");
     });
-    cy.get("#error-message").should("contain", "Invalid login credentials");
-  });
+  })
 });
